Big refactor of animaitons using custom functions
We removed the both the transitions and the gif animations from
`processor/mod.rs`, and but them in their own file. We create
tests to validate the transitions. Finally, we create a function in
`comp_decomp.rs` that allows for packing bytes using customized logic
when finding two different pixels.

This makes it possible for us to create a transition effect simply by
passing a correct closure to that function. It makes the transition
much, much faster, using less memory, and will greatly facilitate
future implementation of new transition effects.