Fix deadlock when sending images quickly
The problem was that we were using SyncSender to stop the animations.

The we use a syncronous channel for sending the frames from the
animation threads in order to ensure old threads that will terminate
soon and new ones don't overlap when sending their frames. However, this
does block the animation threads when they are sending messages.

Because we were also using a syncronous channel to stop the animations,
it was possible that the animation thread was blocked trying to send in
a frame, while the main thread was blocked trying to send in the stop
message.

This fixes that problem by transforming anim_stoppers into a vector of
async Senders.