[client] use scaling when sending images
When the client program is used to send an image request, the image data
is apparently encoded directly into the request.  Now, the sent images
are pre-scaled to the resolution of the pixel buffers used in the
daemon, but the computation of this resolution on the client side was
not accounting for the scale factor.  When the daemon received such
requests, it would attempt to use the given images, and then silently
fail in a worker thread.  This commit simply corrects the client code.