fix integration_tests calling incorrect daemon bin
`tests/integration_tests.rs` was calling the daemon with `swww init`,
which calls the `swww-daemon` that's in the `$PATH`. This means the
tests could fail if the `swww-daemon` under development had changed.
This fixes that by making sure we call the cargo binary instead.

Note that there is a pesky race condition in the tests now, though in
practice I don't expect it to be too much of a problem.