ignore integration_tests by default
This makes testing faster overall (since the integration_tests are very
expensive). The integration_tests only really need to be run once we
finished all of the smaller changes in the program. In order to make
sure we don't forget it, I've put it in the `version.sh` script. The
script will now only succeed if a full testing passed.