Cache animations
swww caches images which takes a really long time for large gifs.
A 160 MB gif takes around 3 minutes to compress.
Taking that long and eating a complete resource is unnecessary because
most switched wallpapers are used more than once.
This commit caches the compressed frames in the `.cache` directory
so that next time the frames can be reused without compression again
(it takes around 3s to load the cache).

Note that the cache is somewhat conservative, if you change some
properties of the animation of the gif it will likely recompress.
This could be improved in the future but I wanted to be sure that no
outdated animations are displayed.