fix gif animations overlapping when sending two gifs
Gif animations were overlapping when an animated gif was sent while
another animated gif was being displayed. This happened because we were
setting `transition-finished` to `true` automatically when the animation
token was dropped. So, when the previous gif animation stopped, it would
drop the token and set `transition-finished` to true, even though the
transition hadn't finished yet.

This patch sets `transition-finished` manually, only for the transitions
that actually finished.