Implement frame callbacks and a new memory pool
We were only using a single WlBuffer per output, busy-waiting until it
was released by the compositor. The problem is that compositors are now
beginning to never release the buffers, keeping them around for some
functionality. This means we have to double-buffer our WlBuffers. Note
that, in order to be future proof, we would like to have something that
automatically triples, or quadruple buffers, as needed.

Now, if we are using two WlBuffers, in order for our animations and
transitions to work, we will need to copy the previous buffer's content
onto the new one at every surface commit.

From my understanding (which could be completely wrong), sctk's current
pools do not let us that. This is because we can't access the content of
any current active buffer without some shenanigans. Therefore, I've
implemented a new memory pool that simply gives us the first free
(non-active) buffer it can find, and allocates more memory automatically
when it can't find any. This means we will automatically double or triple
buffer (or any other number of buffers), as needed.

Now, this leads to a problem: if we don't wait for the compositor, we
will often try to draw before it is ready. If that happens, the buffers
will still be active, and we will have to allocate more. In order to
prevent that, I've implemented proper frame callbacks. This has also had
the benefit of making animations much more smooth, so I suppose this
should have been considered a bug.

There was a problem with the MSRV regarding some structs visibility,
which is why everything has been annotated with pub(super) or
pub(crate).