refactor client's image format choice
We are now sending the image in whatever format the daemon tells us it
is currently supporting, whether that's in 4 channels or 3 channels.
This means we don't need to keep track in the daemon of the image's
format to skip every fourth byte when our WlBuffer_format has an alpha
channel. This will lead to simpler code down the line.

Note that we still make sure to always use 3 channels for animations,
becuase of our compression scheme.