fix transitions performance
Transitions performance took a dip in versions 0.9.* because we were
doing 4x more 'if' statements than necessary when changing the pixels.
This is because some checks runs at every pixel position to see if we
need to change them or not this "round". Our previous refactor had made
it so that these checks were running for every _byte_, not every
_pixel_, which majorly fucked up everything.