fix initial black screen due to late configuration
We were getting a black screen because the initial transitions were
being interrupted, since the compositor would emit a scaling event with
a whole number, followed by a fractional scaling event. The second,
fraction scaling event, would then interrupt transitions since it was a
different scaling factor.

This patch makes it so that, even when a new scaling factor was
advertised, as long as the actual width and height values didn't change,
we do not interrupt transitions.