flush the wayland_fd correctly from animations
event_fd solved a problem where we wouldn't flush the wayland file
descriptor when drawing, and so buffer would never reach the server.

Now, we've allowed for other threads to flush the wayland fd by calling
into a static global. This sidesteps the previous issue without being
necessary to recreate the event_fd.