use just one request for animated images
We had separated them a long time ago in order to send them in a
pipeline. However, that turned out to have the issue of the animation
not starting as soon as possible after the transition was finished, so
we began sending both one after the other in two messages.

That is rather wasteful, so this commit merges everything together in
one big request.