send all draw-related requests in one single msg
This will send attach, damage and frame requests for all wallpapers with
one single request. Later, it will also send all commit requests with
one single request. This is necessary to syncronize all draws when the
user has multiple monitors.

In order to facilitate this, we've turned the code that iterates through
all wallpapers to build the message into its own function. While we were
at it, we did the same for the `stop_animations` function.