Comment swww_randomize_multi.sh and normalize
Added comments to the unusual loop over the displays.
Added a `break 2` to skip to refreshing the list if there are no more images to read.

For the pidfile segment:
Added a read-in of the current process command line according to `/proc/${BASHPID}/comm`. Added OLD_NAME and THIS_NAME variables to make it easier to follow.
Added ${OLD_NAME} to the log message when there's a process with the same PID as the old script instance but with different name. Added a clarification of behavior to the same case.