Fix misuse of read(1) and unset(1)
The use of read(1) in the innermost loop wasn't properly capturing the
piped-in values, and would hang forever waiting for input. The syntax
used in the original swww_randomize_multi.sh does not have this
problem. I don't really understand how the other syntax does not also
have this problem, but I am reverting it so that it at least works.

I removed the $ in front of the $img variable when unsetting it, since
including it was expanding the variable. Oops!