Make build more portable
Ran into this while swww 0.10.0 was being built on the chimera ppc
builder.

NSecs maps to c_long on non-x86_64, which is only guaranteed to be at
least 32 bits (and in fact is often only 32 bits on 32-bit platforms).
While at it, also use Secs for seconds, as while it's always mapped to
i64 currently, it's still more futureproof.