fix animations with 4 channels
Previously, we were using the xrgb shm format, which ignores alpha
values. Because of this, during animations, we could afford to write
garbage every 4th byte, because it would be ignored by the compositor.
Now that we are using the argb shm format, we have to fill up the alpha
channel properly.