fix ghosting when transition-step is high
We were skipping the drawing of the final frame when doing transitions.
This caused problems when the trasition-step is very high, because the
transition would finish midway-through. This effect would be masked when
transition-step was small, which is how I believe most people use
`swww`. It would also be hidden when the images are very different from
each other.