change the way we store the cache
Previously we stored it in `$XDG_CACHE_HOME/swww`.
Now we store it in  `$XDG_CACHE_HOME/swww/$PKG_VERSION`.

This makes it easier for us to eliminate old caches automatically and
allows for easier updates in the future since the user has to do less
things for everything to work.