merge TransitionAnimator and ImageAnimator into one
This merges the TransitionAnimator and ImageAnimator into one Animator
enum. It has the advantage of using less memory than before, because we
always allocated the same amount of TransitionAnimator and ImageAnimator
anyway, and Animator has the same size as TransitionAnimator. It also
cleans up a bit of the code, since we can remove some duplicated things.