refactor to nuke Rc and RefCell from wallpapers
We can use output_name as an unique id, so that we do not need to wrap
everything around `Rc` and `RefCell`.

In some places, this will entail some extra cost in filtering the
`wallpaper` iterator for just those that are part of a group. However,
because the vast majority of our users will probably not have anything
beyong 4 outputs, the cost of this is pretty low. On the other hand,
previously, every access to a wallpaper incurred one extra pointer load
from the `Rc`, and one extra branch for the `RefCell`.

On the other hand, this should improve memory usage and cache access
patterns. This means overall this should also improve performance, or at
the very least not negatively affect it very much.