improve GitHub workflows to only run when necessary
There is no reason to run clippy or `cargo test` when we do not change
Rust code (as in some PRs that improve documentation or example
scripts).

There is no reason to run `cargo deny` if our dependencies did not
change (which we can track by just looking at `Cargo.lock`).

Finally, we added an `cargo_update` action, that will run monthly, and a
`doc-build` action that will run everytime a documentation file changes,
to ensure the docs can always be built.